Docs debugging and directory restructure
