--- projectq/backends/_qrack/.gitignore
@@ -1,1 +1,0 @@
-*.so
