+++ projectq/backends/_sim/_qrack_simulator_test.py
@@ -46,14 +46,14 @@ from projectq.backends import QrackSimulator
 
 
 def test_is_qrack_simulator_present():
-    import projectq.backends._qrack._qracksim
-    assert projectq.backends._qrack._qracksim
+    import projectq.backends._sim._qracksim
+    assert projectq.backends._sim._qracksim
 
 
 def get_available_simulators():
     result = ["py_simulator"]
     try:
-        import projectq.backends._qrack._qracksim as _
+        import projectq.backends._sim._qracksim as _
         result.append("qrack_simulator")
     except ImportError:
         # The Qrack simulator was either not installed or is misconfigured. Skip.
@@ -70,14 +70,9 @@ def get_available_simulators():
 @pytest.fixture(params=get_available_simulators())
 def sim(request):
     if request.param == "qrack_simulator":
-        from projectq.backends._qrack._qracksim import QrackSimulator as QrackSim
+        from projectq.backends._sim._qracksim import QrackSimulator as QrackSim
         sim = QrackSimulator()
         sim._simulator = QrackSim(1)
-        return sim
-    if request.param == "cpp_simulator":
-        from projectq.backends._sim._cppsim import Simulator as CppSim
-        sim = Simulator(gate_fusion=True)
-        sim._simulator = CppSim(1)
         return sim
     if request.param == "py_simulator":
         from projectq.backends._sim._pysim import Simulator as PySim
@@ -515,16 +510,6 @@ class MockSimulatorBackend(object):
 
     def run(self):
         self.run_cnt += 1
-
-
-def test_simulator_flush():
-    sim = Simulator()
-    sim._simulator = MockSimulatorBackend()
-
-    eng = MainEngine(sim)
-    eng.flush()
-
-    assert sim._simulator.run_cnt == 1
 
 
 def test_simulator_functional_entangle(sim):
