+++ setup.py
@@ -54,8 +54,8 @@ qracksim = Feature(
     standard=False,
     ext_modules=[
         Extension(
-            'projectq.backends._qrack._qracksim',
-            ['projectq/backends/_qrack/_qracksim.cpp'],
+            'projectq.backends._sim._qracksim',
+            ['projectq/backends/_sim/_qracksim.cpp'],
             include_dirs=[
                 # Path to pybind11 headers
                 get_pybind_include(),
